requirements.txt
# C:\dev\Captions_Formatter\Formatter_machine\requirements.txt

# Core ISCE dependencies
pyyaml
pandas

# ASR & Alignment Dependencies
pysrt
whisperx
pyannote.audio
ffmpeg-python
rapidfuzz

# SpaCy and its Swedish model
spacy
https://github.com/explosion/spacy-models/releases/download/sv_core_news_lg-3.7.2/sv_core_news_lg-3.7.2-py3-none-any.whl

# Note: Specific versions have been removed to allow pip to resolve
# the latest compatible versions. If you encounter issues, you may need
# to pin specific versions, e.g., pandas==2.1.4